+++ src/reducer/gameData.js
@@ -1,7 +1,7 @@
 export default (state = [], action = {}) => {
     switch (action.type) {
       case "SET_BREEDS": {
-        return action.payload
+        return [...state, action.payload]
       }
       default:
         return state
